+++ modules/dateModule.js
@@ -1,0 +1,12 @@
+const CurrentDate = () => {
+  const dateElement = document.getElementById('date');
+  const currentDate = new Date();
+
+  // Get the current date and time
+  const formattedDate = currentDate.toLocaleString();
+
+  // Set the formatted date as the content of the element
+  dateElement.textContent = formattedDate;
+};
+
+export default CurrentDate;
