+++ modules/elements.js
@@ -1,13 +1,12 @@
-
 const Booklink = document.getElementById('book-link');
 const Addlink = document.getElementById('add-link');
-const Contactlink = document.getElementById('contact-link')
+const Contactlink = document.getElementById('contact-link');
 const Booklist = document.getElementById('booklist');
 const AddBook = document.getElementById('Add-Books');
 const Contact = document.getElementById('contact');
 const main = document.getElementById('hidden');
 
 export {
-    Booklink, Addlink, Contactlink, Booklist, AddBook,
-    Contact, main,
+  Booklink, Addlink, Contactlink, Booklist, AddBook,
+  Contact, main,
 };
